Add msbuild task (#8)
Add a new "build" task source and MSBuild focused build task. We'll eventually add an SDK-style build task as well that works without an MSBuild pre-requisite. The current implementation assumes that MSBuild will already be in the PATH. We can add support for a configurable build path in a follow up change.